temp: tried moving utils to redux folder
- it didn't work
